Extract table name utils
