+++ lib/assets/javascripts/cartodb3/data/layer-definition-model.js
@@ -7,7 +7,7 @@ var DataSQLModel = require('../editor/layers/layer-content-views/data/data-sql-m
 var layerTypesAndKinds = require('./layer-types-and-kinds');
 var InfowindowModel = require('./infowindow-click-model');
 var TooltipModel = require('./infowindow-hover-model');
-var SQLUtils = require('../helpers/sql-utils');
+var TableNameUtils = require('../helpers/table-name-utils');
 
 // from_layer_id and from_letter are not attributes for the model, but are sent to the layer creation
 // endpoint when creating a layer from an existing analysis node (see user-actions)
@@ -235,7 +235,7 @@ module.exports = Backbone.Model.extend({
 
   getQualifiedTableName: function () {
     var userName = this.get('user_name') || this.collection.userModel.get('username');
-    return SQLUtils.getQualifiedTableName(
+    return TableNameUtils.getQualifiedTableName(
       this.getTableName(),
       userName,
       this.collection.userModel.isInsideOrg()
