+++ lib/assets/javascripts/cartodb3/data/table-model.js
@@ -4,6 +4,7 @@ var TableColumnsCollection = require('./table-columns-collection');
 var getSimpleGeometryType = require('./get-simple-geometry-type');
 var PermissionModel = require('./permission-model');
 var SyncModel = require('./synchronization-model');
+var TableNameUtils = require('../helpers/table-name-utils');
 
 /**
  * Model representing a table.
@@ -84,13 +85,7 @@ module.exports = Backbone.Model.extend({
   },
 
   getUnqualifiedName: function () {
-    var name = this.get('name');
-    if (!name) return null;
-    var tk = name.split('.');
-    if (tk.length === 2) {
-      return tk[1];
-    }
-    return name;
+    return TableNameUtils.getUnqualifiedName(this.get('name'));
   },
 
   getOwnerName: function () {
