+++ lib/assets/javascripts/cartodb3/dataset/dataset-content/dataset-content-view.js
@@ -1,6 +1,7 @@
 var CoreView = require('backbone/core-view');
 var TableManager = require('../../components/table/table-manager');
 var DatasetContentOptionsView = require('./dataset-content-options-view');
+var TableNameUtils = require('../../helpers/table-name-utils');
 var SQLUtils = require('../../helpers/sql-utils');
 
 module.exports = CoreView.extend({
@@ -49,7 +50,7 @@ module.exports = CoreView.extend({
     var hasWriteAccess = permissionModel.isOwner(this._userModel) || permissionModel.hasWriteAccess(this._userModel);
     var isReadOnly = this._syncModel.isSync() || !hasWriteAccess;
 
-    var tableName = SQLUtils.getQualifiedTableName(
+    var tableName = TableNameUtils.getQualifiedTableName(
       this._tableModel.getUnqualifiedName(),
       this._tableModel.get('permission').owner.username,
       this._userModel.isInsideOrg()
