+++ lib/assets/javascripts/cartodb3/editor/editor-hints/factory-hints.js
@@ -1,5 +1,5 @@
 var _ = require('underscore');
-var SQLUtils = require('../../helpers/sql-utils');
+var TableNameUtils = require('../../helpers/table-name-utils');
 
 var memo = {};
 
@@ -45,7 +45,7 @@ module.exports = {
 
   _getTableName: function () {
     var tableName = this._layerDefinitionModel.getTableName();
-    tableName = SQLUtils.getUnqualifiedName(tableName);
+    tableName = TableNameUtils.getUnqualifiedName(tableName);
     return [{
       text: tableName,
       type: 'Table_name'
