+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-view.js
@@ -15,7 +15,7 @@ var changeViewButtons = require('./change-view-buttons.tpl');
 var DataView = require('./layer-content-views/data/data-view');
 var LegendsView = require('./layer-content-views/legend/legends-view');
 var FeatureDefinitionModel = require('../../data/feature-definition-model');
-var SQLUtils = require('../../helpers/sql-utils');
+var TableNameUtils = require('../../helpers/table-name-utils');
 
 var REQUIRED_OPTS = [
   'userActions',
@@ -280,8 +280,8 @@ module.exports = CoreView.extend({
 
     // Adds the owner username if it's in an organization and the table name is not already qualified
     var tableName = analysisDefinitionNodeModel.get('table_name');
-    var userName = SQLUtils.getUsernameFromTableName(tableName) || this._userModel.get('username');
-    tableName = SQLUtils.getQualifiedTableName(
+    var userName = TableNameUtils.getUsernameFromTableName(tableName) || this._userModel.get('username');
+    tableName = TableNameUtils.getQualifiedTableName(
       tableName,
       userName,
       this._userModel.isInsideOrg()
