+++ lib/assets/javascripts/cartodb3/helpers/sql-utils.js
@@ -2,6 +2,8 @@
  *  Checks properties from a SQL
  */
 
+var TableNameUtils = require('./table-name-utils');
+
 module.exports = {
   isSameQuery: function (originalQuery, customQuery) {
     if (originalQuery == null || customQuery == null) { // eslint-disable-line
@@ -62,43 +64,7 @@ module.exports = {
     return sql.replace(/\r?\n|\r/g, ' ');
   },
 
-  getUnqualifiedName: function (tablename) {
-    if (!tablename) return null;
-    var tk = tablename.split('.');
-    if (tk.length === 2) {
-      return this.getUnquotedName(tk[1]);
-    }
-    return this.getUnquotedName(tablename);
-  },
-
-  getUsernameFromTableName: function (tablename) {
-    if (!tablename) return null;
-    var tk = tablename.split('.');
-    if (tk.length === 2) {
-      return this.getUnquotedName(tk[0]);
-    }
-    return '';
-  },
-
-  getUnquotedName: function (tablename) {
-    return tablename && tablename.replace(/"/g, '');
-  },
-
-  quoteIfNeeded: function (name) {
-    var VALID_IDENTIFIER = /^[a-zA-Z_][a-zA-Z0-9_$]*$/;
-    name = this.getUnquotedName(name);
-    if (!VALID_IDENTIFIER.test(name)) {
-      name = '"' + name + '"';
-    }
-    return name;
-  },
-
-  getQualifiedTableName: function (tableName, userName, inOrganization) {
-    var schemaPrefix = (inOrganization && userName) ? this.quoteIfNeeded(userName) + '.' : '';
-    return schemaPrefix + this.quoteIfNeeded(this.getUnqualifiedName(tableName));
-  },
-
   getDefaultSQL: function (tableName, userName, inOrganization) {
-    return 'SELECT * FROM ' + this.getQualifiedTableName(tableName, userName, inOrganization);
+    return 'SELECT * FROM ' + TableNameUtils.getQualifiedTableName(tableName, userName, inOrganization);
   }
 };
